requirements.txt
streamlit==1.12.0
rembg==2.0.21
Pillow==8.3.2
werkzeug==2.1.2
pandas==1.3.3
onnxruntime==1.11.1  # Added for rembg's dependency
